fix(numeric): correct the return annotation the new repair key invalidated
collect_page_repairs_by_line still declared dict[tuple[int, int, int], ...] after
group_repairs_by_line started returning a (page, y0, x0) key. Caught by `ty`, which went
from 8 diagnostics to 9 -- established like for like by restoring the five changed files
to their f7dd065 content and running the identical invocation, because ty's count is a
property of the invocation and not of the code. Back to 8.

+++ src/likhit/extractors/numeric_boundaries.py
@@ -401,7 +401,7 @@ def collect_page_repairs_by_line(
     page: object,
     *,
     page_number: int,
-) -> dict[tuple[int, int, int], list[NumericBoundaryRepair]]:
+) -> dict[tuple[int, float, float], list[NumericBoundaryRepair]]:
     """Collect and index one page's repairs, degrading to none on failure.
 
     Both extraction entry points call this rather than pairing the collector
